1.modify statistics function to common tools
